backend/requirements.txt
asgiref==3.10.0
chardet==5.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
google-auth==2.27.0
pypdf==6.1.1
python-dotenv==1.0.1
reportlab==4.2.2
sqlparse==0.5.3
requests==2.32.3
django-redis==5.2.0
python-json-logger==2.0.7
django-prometheus==2.4.0
